+++ babel.config.js
@@ -19,6 +19,7 @@ module.exports = {
           '@contexts': './src/contexts',
           '@navigation': './src/navigation',
           '@storage': './src/storage',
+          '@store': './src/store',
           '@types': './src/types',
         },
       },
