+++ src/store/cards.ts
@@ -1,0 +1,26 @@
+import {useQuery} from 'react-query';
+
+import {sortByDateLastActivity} from '@helpers';
+import {fetchData} from '@network/apiMethods';
+import {getMyCardsUrl} from '@network/apiUrls';
+
+export const useCardsQuery = (token: string, onSuccessHandler?: () => void) => {
+  return useQuery(['cards', token], () => fetchCardsData(token), {
+    enabled: !!token,
+    onSuccess: onSuccessHandler || (() => {}),
+  });
+};
+
+const fetchCardsData = async (token: string) => {
+  const allCardsInfo = await fetchData(getMyCardsUrl(token));
+  const allCards: AllCards = allCardsInfo.map(
+    ({id, idBoard, name, dateLastActivity}: CardInfo) => ({
+      id,
+      idBoard,
+      name,
+      dateLastActivity,
+    }),
+  );
+
+  return allCards.sort(sortByDateLastActivity);
+};
